requirements.txt
azure-storage-blob
azure-cosmos
azure-cognitiveservices-vision-computervision
flask
